feat(api): :sparkles: separate uploadthing to api package so it can be used in nextjs and expo
