+++ apps/nextjs/src/app/api/uploadthing/route.ts
@@ -1,11 +1,9 @@
 import { createRouteHandler } from "uploadthing/next";
 
-import { ourFileRouter } from "./core";
+import { uploadRouter } from "@board-games/api/uploadthing";
 
 // Export routes for Next App Router
 export const { GET, POST } = createRouteHandler({
-  router: ourFileRouter,
-
-  // Apply an (optional) custom config:
-  // config: { ... },
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+  router: uploadRouter,
 });
