+++ packages/api/package.json
@@ -7,7 +7,8 @@
     ".": {
       "types": "./dist/index.d.ts",
       "default": "./src/index.ts"
-    }
+    },
+    "./uploadthing": "./src/uploadthing.ts"
   },
   "license": "MIT",
   "scripts": {
@@ -21,7 +22,9 @@
   "dependencies": {
     "@board-games/db": "workspace:*",
     "@trpc/server": "catalog:",
+    "@uploadthing/react": "^7.1.3",
     "superjson": "2.2.2",
+    "uploadthing": "^7.4.1",
     "zod": "catalog:"
   },
   "devDependencies": {
